# nodejs-server-server-generated/api/swagger.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "AdventureWorksAPI"
paths:
...
  /api/CustomerAddresses:
    get:
      tags:
      - "CustomerAddresses"
      operationId: "GetCustomerAddress"
      consumes: []
      produces:
      - "text/plain"
      - "application/json"
      - "text/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/CustomerAddress"
            uniqueItems: false
      x-swagger-router-controller: "CustomerAddresses"
    post:
      tags:
      - "CustomerAddresses"
      operationId: "PostCustomerAddress"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - in: "body"
        name: "customerAddress"
        required: false
        schema:
          $ref: "#/definitions/CustomerAddress"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "CustomerAddresses"
  /api/CustomerAddresses/{id}:
    get:
      tags:
      - "CustomerAddresses"
      operationId: "GetCustomerAddress"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "CustomerAddresses"
    put:
      tags:
      - "CustomerAddresses"
      operationId: "PutCustomerAddress"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "customerAddress"
        required: false
        schema:
          $ref: "#/definitions/CustomerAddress"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "CustomerAddresses"
    delete:
      tags:
      - "CustomerAddresses"
      operationId: "DeleteCustomerAddress"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "CustomerAddresses"
...
  /api/ErrorLogs:
    get:
      tags:
      - "ErrorLogs"
      operationId: "GetErrorLog"
      consumes: []
      produces:
      - "text/plain"
      - "application/json"
      - "text/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ErrorLog"
            uniqueItems: false
      x-swagger-router-controller: "ErrorLogs"
    post:
      tags:
      - "ErrorLogs"
      operationId: "PostErrorLog"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - in: "body"
        name: "errorLog"
        required: false
        schema:
          $ref: "#/definitions/ErrorLog"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ErrorLogs"
  /api/ErrorLogs/{id}:
    get:
      tags:
      - "ErrorLogs"
      operationId: "GetErrorLog"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ErrorLogs"
    put:
      tags:
      - "ErrorLogs"
      operationId: "PutErrorLog"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "errorLog"
        required: false
        schema:
          $ref: "#/definitions/ErrorLog"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ErrorLogs"
    delete:
      tags:
      - "ErrorLogs"
      operationId: "DeleteErrorLog"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ErrorLogs"
...
  /api/ProductModelProductDescriptions:
    get:
      tags:
      - "ProductModelProductDescriptions"
      operationId: "GetProductModelProductDescription"
      consumes: []
      produces:
      - "text/plain"
      - "application/json"
      - "text/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ProductModelProductDescription"
            uniqueItems: false
      x-swagger-router-controller: "ProductModelProductDescriptions"
    post:
      tags:
      - "ProductModelProductDescriptions"
      operationId: "PostProductModelProductDescription"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - in: "body"
        name: "productModelProductDescription"
        required: false
        schema:
          $ref: "#/definitions/ProductModelProductDescription"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModelProductDescriptions"
  /api/ProductModelProductDescriptions/{id}:
    get:
      tags:
      - "ProductModelProductDescriptions"
      operationId: "GetProductModelProductDescription"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModelProductDescriptions"
    put:
      tags:
      - "ProductModelProductDescriptions"
      operationId: "PutProductModelProductDescription"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "productModelProductDescription"
        required: false
        schema:
          $ref: "#/definitions/ProductModelProductDescription"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModelProductDescriptions"
    delete:
      tags:
      - "ProductModelProductDescriptions"
      operationId: "DeleteProductModelProductDescription"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModelProductDescriptions"
  /api/ProductModels:
    get:
      tags:
      - "ProductModels"
      operationId: "GetProductModel"
      consumes: []
      produces:
      - "text/plain"
      - "application/json"
      - "text/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ProductModel"
            uniqueItems: false
      x-swagger-router-controller: "ProductModels"
    post:
      tags:
      - "ProductModels"
      operationId: "PostProductModel"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - in: "body"
        name: "productModel"
        required: false
        schema:
          $ref: "#/definitions/ProductModel"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModels"
  /api/ProductModels/{id}:
    get:
      tags:
      - "ProductModels"
      operationId: "GetProductModel"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModels"
    put:
      tags:
      - "ProductModels"
      operationId: "PutProductModel"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "productModel"
        required: false
        schema:
          $ref: "#/definitions/ProductModel"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModels"
    delete:
      tags:
      - "ProductModels"
      operationId: "DeleteProductModel"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "ProductModels"
...
  /api/SalesOrderHeaders:
    get:
      tags:
      - "SalesOrderHeaders"
      operationId: "GetSalesOrderHeader"
      consumes: []
      produces:
      - "text/plain"
      - "application/json"
      - "text/json"
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SalesOrderHeader"
            uniqueItems: false
      x-swagger-router-controller: "SalesOrderHeaders"
    post:
      tags:
      - "SalesOrderHeaders"
      operationId: "PostSalesOrderHeader"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - in: "body"
        name: "salesOrderHeader"
        required: false
        schema:
          $ref: "#/definitions/SalesOrderHeader"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "SalesOrderHeaders"
  /api/SalesOrderHeaders/{id}:
    get:
      tags:
      - "SalesOrderHeaders"
      operationId: "GetSalesOrderHeader"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "SalesOrderHeaders"
    put:
      tags:
      - "SalesOrderHeaders"
      operationId: "PutSalesOrderHeader"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      - in: "body"
        name: "salesOrderHeader"
        required: false
        schema:
          $ref: "#/definitions/SalesOrderHeader"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "SalesOrderHeaders"
    delete:
      tags:
      - "SalesOrderHeaders"
      operationId: "DeleteSalesOrderHeader"
      consumes: []
      produces: []
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
        format: "int32"
      responses:
        200:
          description: "Success"
      x-swagger-router-controller: "SalesOrderHeaders"
...
definitions:
...
  SalesOrderHeader:
    type: "object"
    properties:
      salesOrderId:
        type: "integer"
        format: "int32"
      revisionNumber:
        type: "integer"
        format: "int32"
      orderDate:
        type: "string"
        format: "date-time"
      dueDate:
        type: "string"
        format: "date-time"
      shipDate:
        type: "string"
        format: "date-time"
      status:
        type: "integer"
        format: "int32"
      onlineOrderFlag:
        type: "boolean"
      salesOrderNumber:
        type: "string"
      purchaseOrderNumber:
        type: "string"
      accountNumber:
        type: "string"
      customerId:
        type: "integer"
        format: "int32"
      shipToAddressId:
        type: "integer"
        format: "int32"
      billToAddressId:
        type: "integer"
        format: "int32"
      shipMethod:
        type: "string"
      creditCardApprovalCode:
        type: "string"
      subTotal:
        type: "number"
        format: "double"
      taxAmt:
        type: "number"
        format: "double"
      freight:
        type: "number"
        format: "double"
      totalDue:
        type: "number"
        format: "double"
      comment:
        type: "string"
      rowguid:
        type: "string"
        format: "uuid"
      modifiedDate:
        type: "string"
        format: "date-time"
      billToAddress:
        $ref: "#/definitions/Address"
      customer:
        $ref: "#/definitions/Customer"
      shipToAddress:
        $ref: "#/definitions/Address"
      salesOrderDetail:
        type: "array"
        items:
          $ref: "#/definitions/SalesOrderDetail"
    example:
      salesOrderId: 5
      revisionNumber: 2
      dueDate: "2000-01-23T04:56:07.000+00:00"
      freight: 1.231513536777255612975068288506008684635162353515625
      totalDue: 1.024645700144157789424070870154537260532379150390625
      billToAddressId: 2
      subTotal: 4.1456080298839363962315474054776132106781005859375
      shipToAddressId: 3
      rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      customerId: 9
      salesOrderNumber: "salesOrderNumber"
      taxAmt: 7.3862819483858839220147274318151175975799560546875
      onlineOrderFlag: true
      shipDate: "2000-01-23T04:56:07.000+00:00"
      accountNumber: "accountNumber"
      creditCardApprovalCode: "creditCardApprovalCode"
      purchaseOrderNumber: "purchaseOrderNumber"
      shipMethod: "shipMethod"
      modifiedDate: "2000-01-23T04:56:07.000+00:00"
      comment: "comment"
      salesOrderDetail:
      - salesOrderId: 1
        unitPrice: 4.9652184929849543237878606305457651615142822265625
        product:
          productModel:
            product:
            - null
            - null
            name: "name"
            modifiedDate: "2000-01-23T04:56:07.000+00:00"
            productModelId: 2
            rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
            productModelProductDescription:
            - culture: "culture"
              productDescriptionId: 6
              modifiedDate: "2000-01-23T04:56:07.000+00:00"
              productModelId: 6
              rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
              productDescription:
                productDescriptionId: 5
                modifiedDate: "2000-01-23T04:56:07.000+00:00"
                description: "description"
                rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
                productModelProductDescription:
                - null
                - null
            - culture: "culture"
              productDescriptionId: 6
              modifiedDate: "2000-01-23T04:56:07.000+00:00"
              productModelId: 6
              rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
              productDescription:
                productDescriptionId: 5
                modifiedDate: "2000-01-23T04:56:07.000+00:00"
                description: "description"
                rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
                productModelProductDescription:
                - null
                - null
            catalogDescription: "catalogDescription"
          thumbNailPhoto: "thumbNailPhoto"
          productId: 9
          color: "color"
          weight: 9.0183481860707832566959041287191212177276611328125
          productNumber: "productNumber"
          thumbnailPhotoFileName: "thumbnailPhotoFileName"
          rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          standardCost: 6.683562403749608193948006373830139636993408203125
          sellEndDate: "2000-01-23T04:56:07.000+00:00"
          productCategory:
            product:
            - null
            - null
            productCategoryId: 6
            name: "name"
            modifiedDate: "2000-01-23T04:56:07.000+00:00"
            parentProductCategoryId: 1
            inverseParentProductCategory:
            - null
            - null
            rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          productCategoryId: 6
          size: "size"
          sellStartDate: "2000-01-23T04:56:07.000+00:00"
          name: "name"
          modifiedDate: "2000-01-23T04:56:07.000+00:00"
          salesOrderDetail:
          - null
          - null
          productModelId: 3
          discontinuedDate: "2000-01-23T04:56:07.000+00:00"
          listPrice: 8.7620420127490010742121739895083010196685791015625
        productId: 1
        unitPriceDiscount: 5.02500479152029466689555192715488374233245849609375
        lineTotal: 9.965781217890562260208753286860883235931396484375
        modifiedDate: "2000-01-23T04:56:07.000+00:00"
        orderQty: 7
        salesOrderDetailId: 6
        rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      - salesOrderId: 1
        unitPrice: 4.9652184929849543237878606305457651615142822265625
        product:
          productModel:
            product:
            - null
            - null
            name: "name"
            modifiedDate: "2000-01-23T04:56:07.000+00:00"
            productModelId: 2
            rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
            productModelProductDescription:
            - culture: "culture"
              productDescriptionId: 6
              modifiedDate: "2000-01-23T04:56:07.000+00:00"
              productModelId: 6
              rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
              productDescription:
                productDescriptionId: 5
                modifiedDate: "2000-01-23T04:56:07.000+00:00"
                description: "description"
                rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
                productModelProductDescription:
                - null
                - null
            - culture: "culture"
              productDescriptionId: 6
              modifiedDate: "2000-01-23T04:56:07.000+00:00"
              productModelId: 6
              rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
              productDescription:
                productDescriptionId: 5
                modifiedDate: "2000-01-23T04:56:07.000+00:00"
                description: "description"
                rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
                productModelProductDescription:
                - null
                - null
            catalogDescription: "catalogDescription"
          thumbNailPhoto: "thumbNailPhoto"
          productId: 9
          color: "color"
          weight: 9.0183481860707832566959041287191212177276611328125
          productNumber: "productNumber"
          thumbnailPhotoFileName: "thumbnailPhotoFileName"
          rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          standardCost: 6.683562403749608193948006373830139636993408203125
          sellEndDate: "2000-01-23T04:56:07.000+00:00"
          productCategory:
            product:
            - null
            - null
            productCategoryId: 6
            name: "name"
            modifiedDate: "2000-01-23T04:56:07.000+00:00"
            parentProductCategoryId: 1
            inverseParentProductCategory:
            - null
            - null
            rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          productCategoryId: 6
          size: "size"
          sellStartDate: "2000-01-23T04:56:07.000+00:00"
          name: "name"
          modifiedDate: "2000-01-23T04:56:07.000+00:00"
          salesOrderDetail:
          - null
          - null
          productModelId: 3
          discontinuedDate: "2000-01-23T04:56:07.000+00:00"
          listPrice: 8.7620420127490010742121739895083010196685791015625
        productId: 1
        unitPriceDiscount: 5.02500479152029466689555192715488374233245849609375
        lineTotal: 9.965781217890562260208753286860883235931396484375
        modifiedDate: "2000-01-23T04:56:07.000+00:00"
        orderQty: 7
        salesOrderDetailId: 6
        rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      orderDate: "2000-01-23T04:56:07.000+00:00"
      status: 7
...
  ProductModel:
    type: "object"
    properties:
      productModelId:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      catalogDescription:
        type: "string"
      rowguid:
        type: "string"
        format: "uuid"
      modifiedDate:
        type: "string"
        format: "date-time"
      product:
        type: "array"
        items:
          $ref: "#/definitions/Product"
      productModelProductDescription:
        type: "array"
        items:
          $ref: "#/definitions/ProductModelProductDescription"
    example:
      product:
      - null
      - null
      name: "name"
      modifiedDate: "2000-01-23T04:56:07.000+00:00"
      productModelId: 2
      rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      productModelProductDescription:
      - culture: "culture"
        productDescriptionId: 6
        modifiedDate: "2000-01-23T04:56:07.000+00:00"
        productModelId: 6
        rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        productDescription:
          productDescriptionId: 5
          modifiedDate: "2000-01-23T04:56:07.000+00:00"
          description: "description"
          rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          productModelProductDescription:
          - null
          - null
      - culture: "culture"
        productDescriptionId: 6
        modifiedDate: "2000-01-23T04:56:07.000+00:00"
        productModelId: 6
        rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        productDescription:
          productDescriptionId: 5
          modifiedDate: "2000-01-23T04:56:07.000+00:00"
          description: "description"
          rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
          productModelProductDescription:
          - null
          - null
      catalogDescription: "catalogDescription"
  ProductModelProductDescription:
    type: "object"
    properties:
      productModelId:
        type: "integer"
        format: "int32"
      productDescriptionId:
        type: "integer"
        format: "int32"
      culture:
        type: "string"
      rowguid:
        type: "string"
        format: "uuid"
      modifiedDate:
        type: "string"
        format: "date-time"
      productDescription:
        $ref: "#/definitions/ProductDescription"
      productModel:
        $ref: "#/definitions/ProductModel"
    example:
      culture: "culture"
      productDescriptionId: 6
      modifiedDate: "2000-01-23T04:56:07.000+00:00"
      productModelId: 6
      rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      productDescription:
        productDescriptionId: 5
        modifiedDate: "2000-01-23T04:56:07.000+00:00"
        description: "description"
        rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
        productModelProductDescription:
        - null
        - null
  ProductDescription:
    type: "object"
    properties:
      productDescriptionId:
        type: "integer"
        format: "int32"
      description:
        type: "string"
      rowguid:
        type: "string"
        format: "uuid"
      modifiedDate:
        type: "string"
        format: "date-time"
      productModelProductDescription:
        type: "array"
        items:
          $ref: "#/definitions/ProductModelProductDescription"
    example:
      productDescriptionId: 5
      modifiedDate: "2000-01-23T04:56:07.000+00:00"
      description: "description"
      rowguid: "046b6c7f-0b8a-43b9-b35d-6489e6daee91"
      productModelProductDescription:
      - null
      - null
  ErrorLog:
    type: "object"
    properties:
      errorLogId:
        type: "integer"
        format: "int32"
      errorTime:
        type: "string"
        format: "date-time"
      errorNumber:
        type: "integer"
        format: "int32"
      errorSeverity:
        type: "integer"
        format: "int32"
      errorState:
        type: "integer"
        format: "int32"
      errorProcedure:
        type: "string"
      errorLine:
        type: "integer"
        format: "int32"
      errorMessage:
        type: "string"
    example:
      errorTime: "2000-01-23T04:56:07.000+00:00"
      errorState: 5
      errorProcedure: "errorProcedure"
      errorSeverity: 1
      errorMessage: "errorMessage"
      errorLine: 5
      errorLogId: 0
      errorNumber: 6