# rails5-server-generated/app/models/product_description.rb
=begin
AdventureWorksAPI

No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)

OpenAPI spec version: v1

Generated by: https://github.com/swagger-api/swagger-codegen.git

=end


class ProductDescription < ApplicationRecord

  serialize :product_model_product_description, Array
end